split state transition test
